requirements.txt
camoufox
curl_cffi
Flask
Pillow
Requests
websocket_client